Organize orbit.envelope module
Separate files:
- envelopy.py has Envelope class definition.
- track.py has EnvelopeTracker class definition.
- matrix.py has functions to get transfer matrices from AccNode
