+++ py/orbit/envelope/__init__.py
@@ -1,3 +1,2 @@
 from .envelope import Envelope
-from .envelope import EnvelopeTracker
-from . import matrix
+from .track import EnvelopeTracker
