+++ py/orbit/envelope/envelope.py
@@ -1,75 +1,13 @@
-import math
-import warnings
-
 import numpy as np
 import scipy.constants
 import scipy.special
 
 from orbit.core.bunch import Bunch
 from orbit.core.bunch import SyncParticle
-from orbit.lattice import AccNode
-from orbit.lattice import AccLattice
 
-from orbit.teapot import ApertureTEAPOT
-from orbit.teapot import DriftTEAPOT
-from orbit.teapot import BendTEAPOT
-from orbit.teapot import KickTEAPOT
-from orbit.teapot import MonitorTEAPOT
-from orbit.teapot import MultipoleTEAPOT
-from orbit.teapot import NodeTEAPOT
-from orbit.teapot import QuadTEAPOT
-from orbit.teapot import SolenoidTEAPOT
-from orbit.teapot import FringeFieldTEAPOT
-from orbit.teapot import BunchWrapTEAPOT
-from orbit.teapot import TiltTEAPOT
-from orbit.teapot import ContinuousLinearFocusingTEAPOT
-from orbit.teapot import TurnCounterTEAPOT
-
-from orbit.py_linac.lattice import MarkerLinacNode as MarkerLINAC
-from orbit.py_linac.lattice import Drift as DriftLINAC
-from orbit.py_linac.lattice import Quad as QuadLINAC
-from orbit.py_linac.lattice import Bend as BendLINAC
-from orbit.py_linac.lattice import DCorrectorH as DCorrectorHLINAC
-from orbit.py_linac.lattice import DCorrectorV as DCorrectorVLINAC
-from orbit.py_linac.lattice import Solenoid as SolenoidLINAC
-from orbit.py_linac.lattice import TiltElement as TiltLINAC
-from orbit.py_linac.lattice import FringeField as FringeFieldLINAC
-from orbit.py_linac.lattice import BaseRF_Gap as BaseRF_Gap
-from orbit.py_linac.lattice import LinacApertureNode as ApertureLINAC
-
-from .matrix import get_dp_p_coeff
-from .matrix import get_zp_coeff
-from .matrix import convert_matrix_dp_p_to_dE
 from .matrix import convert_matrix_zp_to_dE
-from .matrix import get_matrix_tilt
-from .matrix import get_matrix_kick
-from .matrix import get_matrix_drift
-from .matrix import get_matrix_quad
-from .matrix import get_matrix_bend
-from .matrix import get_matrix_solenoid
-from .matrix import get_matrix_rf_gap
-from .matrix import get_matrix_cf
 from .utils import gen_dist
 from .utils import proj_cov_matrix
-
-
-ENTRANCE = AccNode.ENTRANCE
-BODY = AccNode.BODY
-EXIT = AccNode.EXIT
-
-BEFORE = AccNode.BEFORE
-AFTER = AccNode.AFTER
-
-IGNORE_NODE_TYPES = [
-    NodeTEAPOT,
-    MonitorTEAPOT,
-    FringeFieldTEAPOT,
-    ApertureTEAPOT,
-    BunchWrapTEAPOT,
-    TurnCounterTEAPOT,
-    MarkerLINAC,
-    FringeFieldLINAC,
-]
 
 
 def build_diag_matrix_from_xyz_eig(eigenvectors: np.ndarray) -> np.ndarray:
@@ -80,189 +18,6 @@ def build_diag_matrix_from_xyz_eig(eigenvectors: np.ndarray) -> np.ndarray:
             col = j * 2
             A[row, col] = A[row + 1, col + 1] = eigenvectors[i, j]
     return A
-
-
-def get_matrix(
-    node: AccNode,
-    sync_part: SyncParticle,
-    charge: float,
-    index: int = -1,
-) -> np.ndarray | None:
-    """Calculate transfer matrix and update synchronous particle.
-
-    This function maps various accelerator nodes to 7 x 7 transfer matrices
-    for envelope tracking. For non-accelerating, finite-length nodes, the
-    synchronous particle time is updated as in a drift. Accelerating nodes
-    such as RF gaps will update the synchronous particle energy.
-
-    Args:
-        node: The accelerator node.
-        sync_part: Synchronous particle.
-        charge: Particle charge. (The charge is currently an attribute of the
-            bunch, not the synchronous particle.)
-        index: Node part index. An index of -1 will return the transfer matrix
-            for the entire node.
-    Returns:
-        7 x 7 transfer matrix or None. If None, the node can be ignored during
-        envelope tracking.
-    """
-
-    node_type = type(node)
-    if node_type in IGNORE_NODE_TYPES:
-        return None
-
-    length = node.getLength(index)
-    nparts = node.getnParts()
-
-    if node_type is DriftTEAPOT:
-        if length <= 0:
-            return None
-        return get_matrix_drift(sync_part=sync_part, length=length)
-
-    elif node_type is SolenoidTEAPOT:
-        if length <= 0:
-            return None
-        B = node.getParam("B")
-        if node.waveform:
-            B *= node.waveform.getStrength()
-        return get_matrix_solenoid(sync_part=sync_part, length=length, B=B, charge=charge)
-
-    elif node_type is MultipoleTEAPOT:
-        if length <= 0:
-            return None
-        if np.all(np.abs(node.getParam("kls")) == 0):
-            return get_matrix_drift(sync_part=sync_part, length=length)
-
-    elif node_type is QuadTEAPOT:
-        if length <= 0:
-            return None
-        kq = node.getParam("kq")
-        if node.waveform:
-            kq *= node.waveform.getStrength()
-        return get_matrix_quad(sync_part=sync_part, length=length, kq=kq, charge=charge)
-
-    elif node_type is BendTEAPOT:
-        if length <= 0:
-            return None
-        theta = node.getParam("theta") / (nparts - 1)
-        if index == 0 or index == nparts - 1:
-            theta *= 0.5
-        return get_matrix_bend(sync_part=sync_part, length=length, theta=theta, charge=charge)
-
-    elif node_type is KickTEAPOT:
-        scale = 1.0
-        if node.waveform is not None:
-            scale = node.waveform.getStrength()
-
-        scale /= (nparts - 1)
-        kx = scale * node.getParam("kx")
-        ky = scale * node.getParam("ky")
-        kE = node.getParam("dE")
-
-        if abs(kx) > 0 or abs(ky) > 0 or abs(kE) > 0:
-            return np.matmul(
-                get_matrix_kick(sync_part=sync_part, kx=kx, ky=ky, kE=kE),
-                get_matrix_drift(sync_part=sync_part, length=length),
-            )
-        else:
-            return get_matrix_drift(sync_part=sync_part, length=length)
-
-    elif node_type is TiltTEAPOT:
-        angle = node.getTiltAngle()
-        if angle == 0:
-            return None
-        return get_matrix_tilt(sync_part=sync_part, angle=angle)
-
-    elif node_type is ContinuousLinearFocusingTEAPOT:
-        if length <= 0:
-            return None
-        kq = node.getParam("kq")
-        if node.waveform:
-            kq *= node.waveform.getStrength()
-        return get_matrix_cf(sync_part=sync_part, length=length, kq=kq)
-
-    elif node_type is DriftLINAC:
-        if length <= 0:
-            return None
-        return get_matrix_drift(sync_part=sync_part, length=length)
-
-    elif node_type is QuadLINAC:
-        if length <= 0:
-            return None
-        brho = 3.335640952 * sync_part.momentum() / charge
-        kq = node.getParam("dB/dr") / brho
-        return get_matrix_quad(sync_part=sync_part, length=length, kq=kq, charge=charge)
-
-    elif node_type is BendLINAC:
-        if length <= 0:
-            return None
-        theta = node.getParam("theta") / (nparts - 1)
-        if index == 0 or index == nparts - 1:
-            theta *= 0.5
-        return get_matrix_bend(sync_part=sync_part, length=length, theta=theta, charge=charge)
-
-    elif node_type is DCorrectorHLINAC:
-        length = node.getParam("effLength") / nparts
-        field = node.getParam("B")
-        delta_xp = -field * charge * length * 0.299792 / sync_part.momentum()
-        if delta_xp == 0:
-            return None
-        return get_matrix_kick(sync_part=sync_part, kx=delta_xp, ky=0.0, kE=0.0)
-
-    elif node_type is DCorrectorVLINAC:
-        length = node.getParam("effLength") / nparts
-        field = node.getParam("B")
-        delta_yp = -field * charge * length * 0.299792 / sync_part.momentum()
-        if delta_yp == 0:
-            return None
-        return get_matrix_kick(sync_part=sync_part, kx=0.0, ky=delta_yp, kE=0.0)
-
-    elif node_type is SolenoidLINAC:
-        if length <= 0:
-            return None
-        B = node.getParam("B")
-        return get_matrix_solenoid(sync_part=sync_part, length=length, B=B, charge=charge)
-
-    elif node_type is TiltLINAC:
-        angle = node.getTiltAngle()
-        if angle == 0:
-            return None
-        return get_matrix_tilt(sync_part=sync_part, angle=angle)
-
-    elif node_type is BaseRF_Gap:
-        E0TL = node.getParam("E0TL")
-        mode_phase = node.getParam("mode") * math.pi
-
-        cavity = node.getRF_Cavity()
-        frequency = cavity.getFrequency()
-        phase = cavity.getPhase() + mode_phase
-        amplitude = cavity.getAmp()
-
-        arrival_time = sync_part.time()
-        arrival_time_design = cavity.getDesignArrivalTime()
-
-        if node.isFirstRFGap():
-            if cavity.isDesignSetUp():
-                phase = math.fmod(frequency * (arrival_time - arrival_time_design) * 2.0 * math.pi + phase, 2.0 * math.pi)
-            else:
-                orbitFinalize("Run `trackDesign` first to initialize cavity phases.")
-        else:
-            phase = math.fmod(frequency * (arrival_time - arrival_time_design) * 2.0 * math.pi + phase,2.0 * math.pi)
-
-        node.setGapPhase(phase)
-
-        if amplitude == 0.0:
-            return None
-
-        return get_matrix_rf_gap(
-            sync_part=sync_part,
-            frequency=frequency,
-            E0TL=(E0TL * amplitude),
-            phase=phase,
-            charge=charge,
-        )
-
-    raise NotImplementedError(str(node))
 
 
 class Envelope:
@@ -283,7 +38,7 @@ class Envelope:
         intensity: float = 0.0,
     ) -> None:
 
-        # Eventually allow:
+        # [TO DO]
         #   - setting covariance matrix from bunch particles
         #   - tracking bunch particles as test particles
         empty_bunch = Bunch()
@@ -464,219 +219,3 @@ class Envelope:
 
         # Convert from z' to dE
         return convert_matrix_zp_to_dE(M, self.sync_part)
-
-
-class EnvelopeTracker:
-    def __init__(self, lattice: AccLattice, space_charge: str | None = None) -> None:
-        """Constructor.
-
-        Args:
-            lattice: The accelerator lattice.
-            space_charge: Space charge model {"2d", "3d", None}.
-        """
-        self.lattice = lattice
-        self.space_charge = space_charge
-
-        # For pre-computing elements
-        self.elements = []
-        self.one_turn_matrix = None
-        # [TO DO] option to return one-turn matrix including linear space charge
-
-        for node in self.lattice.getNodes():
-            if type(node) in (BendTEAPOT, BendLINAC):
-                if node.getParam("ea1") != 0.0 or node.getParam("ea2") != 0.0:
-                    message = f"Found bend ea1 or ea2 != 0.0 ({node.getName()}.)"
-                    message += " Nonzero edge angles are not yet supported in envelope tracking."
-                    message += " Setting ea1 and ea2 to 0.0."
-                    warnings.warn(message)
-
-                    node.setParam("ea1", 0.0)
-                    node.setParam("ea2", 0.0)
-
-    def track(self, envelope: Envelope) -> None:
-        sync_part = envelope.sync_part
-        charge = envelope.charge()
-
-        for node_index, node in enumerate(self.lattice.getNodes()):
-            for child_node in node.getChildNodes(ENTRANCE):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    envelope.transform(matrix)
-
-            for part_index in range(node.getnParts()):
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=BEFORE):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        envelope.transform(matrix)
-
-                matrix_sc = None
-                if self.space_charge:
-                    length = node.getLength(part_index)
-                    if length > 0:
-                        if self.space_charge == "2d":
-                            matrix_sc = envelope.sc_matrix_2d(length)
-                        elif self.space_charge == "3d":
-                            matrix_sc = envelope.sc_matrix_3d(length)
-                        else:
-                            raise ValueError
-
-                matrix = get_matrix(node, sync_part=sync_part, charge=charge, index=part_index)
-                if matrix is not None:
-                    if matrix_sc is not None:
-                        matrix = matrix @ matrix_sc
-                    envelope.transform(matrix)
-
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=AFTER):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        envelope.transform(matrix)
-
-            for child_node in node.getChildNodes(EXIT):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    envelope.transform(matrix)
-
-    def track_history(self, envelope: Envelope) -> dict[str, list]:
-        """Track and return envelope parameters vs. position in lattice."""
-        history = {}
-        history["position"] = []
-        history["rms_x"] = []
-        history["rms_y"] = []
-        history["rms_z"] = []
-        history["kin_energy"] = []
-
-        sync_part = envelope.sync_part
-        charge = envelope.charge()
-        node_positions = self.lattice.getNodePositionsDict()
-
-        history["position"].append(0.0)
-        history["rms_x"].append(1000.0 * envelope.rms(0))
-        history["rms_y"].append(1000.0 * envelope.rms(2))
-        history["rms_z"].append(1000.0 * envelope.rms(4))
-        history["kin_energy"].append(envelope.sync_part.kinEnergy())
-
-        for node_index, node in enumerate(self.lattice.getNodes()):
-            for child_node in node.getChildNodes(ENTRANCE):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    envelope.transform(matrix)
-
-            for part_index in range(node.getnParts()):
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=BEFORE):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        envelope.transform(matrix)
-
-                matrix_sc = None
-                if self.space_charge:
-                    length = node.getLength(part_index)
-                    if length > 0:
-                        if self.space_charge == "2d":
-                            matrix_sc = envelope.sc_matrix_2d(length)
-                        elif self.space_charge == "3d":
-                            matrix_sc = envelope.sc_matrix_3d(length)
-                        else:
-                            raise ValueError
-
-                matrix = get_matrix(node, sync_part=sync_part, charge=charge, index=part_index)
-                if matrix is not None:
-                    if matrix_sc is not None:
-                        matrix = matrix @ matrix_sc
-                    envelope.transform(matrix)
-
-                position_start, position_stop = node_positions[node]
-                position = position_start + node.getLength(part_index) * (part_index + 1)
-
-                history["position"].append(position)
-                history["rms_x"].append(1000.0 * envelope.rms(0))
-                history["rms_y"].append(1000.0 * envelope.rms(2))
-                history["rms_z"].append(1000.0 * envelope.rms(4))
-                history["kin_energy"].append(envelope.sync_part.kinEnergy())
-
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=AFTER):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        envelope.transform(matrix)
-
-            for child_node in node.getChildNodes(EXIT):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    envelope.transform(matrix)
-
-        return history
-
-    def precompute_matrices(self, envelope: Envelope) -> None:
-        """Pre-compute transfer matrices for each node.
-
-        For each node, return tuple (node, matrix). Mark space charge kicks as ("sc", length).
-        """
-        sync_part = envelope.sync_part
-        charge = envelope.charge()
-
-        self.elements = []
-        for node_index, node in enumerate(self.lattice.getNodes()):
-            for child_node in node.getChildNodes(ENTRANCE):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    self.elements.append((child_node, matrix))
-
-            for part_index in range(node.getnParts()):
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=BEFORE):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        self.elements.append((child_node, matrix))
-
-                if self.space_charge:
-                    length = node.getLength(part_index)
-                    if length > 0:
-                        self.elements.append(("sc", length))
-
-                matrix = get_matrix(node, sync_part=sync_part, charge=charge, index=part_index)
-                if matrix is not None:
-                    self.elements.append((node, matrix))
-
-                for child_node in node.getChildNodes(BODY, part_index, place_in_part=AFTER):
-                    matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                    if matrix is not None:
-                        self.elements.append((node, matrix))
-
-            for child_node in node.getChildNodes(EXIT):
-                matrix = get_matrix(child_node, sync_part=sync_part, charge=charge)
-                if matrix is not None:
-                    self.elements.append((node, matrix))
-
-    def track_ring(self, envelope: Envelope) -> None:
-        """Track using pre-computed transfer matrices.
-
-        The method assumes that all nodes are static and that there is no
-        change in the synchronous particle energy. In this case the matrices
-        can be computed once and reused on each turn. If there is no space charge,
-        we track using the one-turn matrix.
-        """
-
-        # Pre-compute transfer matrices on the first turn.
-        if not self.elements:
-            self.precompute_matrices(envelope)
-            self.one_turn_matrix = None
-
-        # If there is no space charge, apply the one-turn transfer matrix.
-        if not self.space_charge:
-            if self.one_turn_matrix is None:
-                self.one_turn_matrix = np.identity(7)
-                for (node, matrix) in self.elements:
-                    self.one_turn_matrix = np.matmul(self.one_turn_matrix, matrix)
-            return envelope.transform(self.one_turn_matrix)
-
-        # If there is space charge, apply the matrices one-by-one.
-        for element in self.elements:
-            if element[0] == "sc":
-                length = element[1]
-                if self.space_charge == "2d":
-                    envelope.transform(envelope.sc_matrix_2d(length))
-                elif self.space_charge == "3d":
-                    envelope.transform(envelope.sc_matrix_3d(length))
-                else:
-                    raise ValueError
-            else:
-                node, matrix = element
-                envelope.transform(matrix)
