+++ py/orbit/envelope/matrix.py
@@ -9,7 +9,45 @@ import numpy as np
 
 from orbit.core.bunch import Bunch
 from orbit.core.bunch import SyncParticle
+from orbit.lattice import AccNode
+from orbit.teapot import ApertureTEAPOT
+from orbit.teapot import DriftTEAPOT
+from orbit.teapot import BendTEAPOT
+from orbit.teapot import KickTEAPOT
+from orbit.teapot import MonitorTEAPOT
+from orbit.teapot import MultipoleTEAPOT
+from orbit.teapot import NodeTEAPOT
+from orbit.teapot import QuadTEAPOT
+from orbit.teapot import SolenoidTEAPOT
+from orbit.teapot import FringeFieldTEAPOT
+from orbit.teapot import BunchWrapTEAPOT
+from orbit.teapot import TiltTEAPOT
+from orbit.teapot import ContinuousLinearFocusingTEAPOT
+from orbit.teapot import TurnCounterTEAPOT
+from orbit.py_linac.lattice import MarkerLinacNode as MarkerLINAC
+from orbit.py_linac.lattice import Drift as DriftLINAC
+from orbit.py_linac.lattice import Quad as QuadLINAC
+from orbit.py_linac.lattice import Bend as BendLINAC
+from orbit.py_linac.lattice import DCorrectorH as DCorrectorHLINAC
+from orbit.py_linac.lattice import DCorrectorV as DCorrectorVLINAC
+from orbit.py_linac.lattice import Solenoid as SolenoidLINAC
+from orbit.py_linac.lattice import TiltElement as TiltLINAC
+from orbit.py_linac.lattice import FringeField as FringeFieldLINAC
+from orbit.py_linac.lattice import BaseRF_Gap as BaseRF_Gap
+from orbit.py_linac.lattice import LinacApertureNode as ApertureLINAC
 from orbit.utils.consts import speed_of_light
+
+
+IGNORE_NODE_TYPES = [
+    NodeTEAPOT,
+    MonitorTEAPOT,
+    FringeFieldTEAPOT,
+    ApertureTEAPOT,
+    BunchWrapTEAPOT,
+    TurnCounterTEAPOT,
+    MarkerLINAC,
+    FringeFieldLINAC,
+]
 
 
 def get_dp_p_coeff(sync_part: SyncParticle) -> float:
@@ -246,3 +284,186 @@ def get_matrix_rf_gap(sync_part: SyncParticle, frequency: float, E0TL: float, ph
     M[1, 0] = d_rp
     M[3, 2] = d_rp
     return M
+
+
+def get_matrix(
+    node: AccNode,
+    sync_part: SyncParticle,
+    charge: float,
+    index: int = -1,
+) -> np.ndarray | None:
+    """Calculate transfer matrix and update synchronous particle.
+
+    This function maps various accelerator nodes to 7 x 7 transfer matrices
+    for envelope tracking. For non-accelerating, finite-length nodes, the
+    synchronous particle time is updated as in a drift. Accelerating nodes
+    such as RF gaps will update the synchronous particle energy.
+
+    Args:
+        node: The accelerator node.
+        sync_part: Synchronous particle.
+        charge: Particle charge. (The charge is currently an attribute of the
+            bunch, not the synchronous particle.)
+        index: Node part index. An index of -1 will return the transfer matrix
+            for the entire node.
+    Returns:
+        7 x 7 transfer matrix or None. If None, the node can be ignored during
+        envelope tracking.
+    """
+
+    node_type = type(node)
+    if node_type in IGNORE_NODE_TYPES:
+        return None
+
+    length = node.getLength(index)
+    nparts = node.getnParts()
+
+    if node_type is DriftTEAPOT:
+        if length <= 0:
+            return None
+        return get_matrix_drift(sync_part=sync_part, length=length)
+
+    elif node_type is SolenoidTEAPOT:
+        if length <= 0:
+            return None
+        B = node.getParam("B")
+        if node.waveform:
+            B *= node.waveform.getStrength()
+        return get_matrix_solenoid(sync_part=sync_part, length=length, B=B, charge=charge)
+
+    elif node_type is MultipoleTEAPOT:
+        if length <= 0:
+            return None
+        if np.all(np.abs(node.getParam("kls")) == 0):
+            return get_matrix_drift(sync_part=sync_part, length=length)
+
+    elif node_type is QuadTEAPOT:
+        if length <= 0:
+            return None
+        kq = node.getParam("kq")
+        if node.waveform:
+            kq *= node.waveform.getStrength()
+        return get_matrix_quad(sync_part=sync_part, length=length, kq=kq, charge=charge)
+
+    elif node_type is BendTEAPOT:
+        if length <= 0:
+            return None
+        theta = node.getParam("theta") / (nparts - 1)
+        if index == 0 or index == nparts - 1:
+            theta *= 0.5
+        return get_matrix_bend(sync_part=sync_part, length=length, theta=theta, charge=charge)
+
+    elif node_type is KickTEAPOT:
+        scale = 1.0
+        if node.waveform is not None:
+            scale = node.waveform.getStrength()
+
+        scale /= (nparts - 1)
+        kx = scale * node.getParam("kx")
+        ky = scale * node.getParam("ky")
+        kE = node.getParam("dE")
+
+        if abs(kx) > 0 or abs(ky) > 0 or abs(kE) > 0:
+            return np.matmul(
+                get_matrix_kick(sync_part=sync_part, kx=kx, ky=ky, kE=kE),
+                get_matrix_drift(sync_part=sync_part, length=length),
+            )
+        else:
+            return get_matrix_drift(sync_part=sync_part, length=length)
+
+    elif node_type is TiltTEAPOT:
+        angle = node.getTiltAngle()
+        if angle == 0:
+            return None
+        return get_matrix_tilt(sync_part=sync_part, angle=angle)
+
+    elif node_type is ContinuousLinearFocusingTEAPOT:
+        if length <= 0:
+            return None
+        kq = node.getParam("kq")
+        if node.waveform:
+            kq *= node.waveform.getStrength()
+        return get_matrix_cf(sync_part=sync_part, length=length, kq=kq)
+
+    elif node_type is DriftLINAC:
+        if length <= 0:
+            return None
+        return get_matrix_drift(sync_part=sync_part, length=length)
+
+    elif node_type is QuadLINAC:
+        if length <= 0:
+            return None
+        brho = 3.335640952 * sync_part.momentum() / charge
+        kq = node.getParam("dB/dr") / brho
+        return get_matrix_quad(sync_part=sync_part, length=length, kq=kq, charge=charge)
+
+    elif node_type is BendLINAC:
+        if length <= 0:
+            return None
+        theta = node.getParam("theta") / (nparts - 1)
+        if index == 0 or index == nparts - 1:
+            theta *= 0.5
+        return get_matrix_bend(sync_part=sync_part, length=length, theta=theta, charge=charge)
+
+    elif node_type is DCorrectorHLINAC:
+        length = node.getParam("effLength") / nparts
+        field = node.getParam("B")
+        delta_xp = -field * charge * length * 0.299792 / sync_part.momentum()
+        if delta_xp == 0:
+            return None
+        return get_matrix_kick(sync_part=sync_part, kx=delta_xp, ky=0.0, kE=0.0)
+
+    elif node_type is DCorrectorVLINAC:
+        length = node.getParam("effLength") / nparts
+        field = node.getParam("B")
+        delta_yp = -field * charge * length * 0.299792 / sync_part.momentum()
+        if delta_yp == 0:
+            return None
+        return get_matrix_kick(sync_part=sync_part, kx=0.0, ky=delta_yp, kE=0.0)
+
+    elif node_type is SolenoidLINAC:
+        if length <= 0:
+            return None
+        B = node.getParam("B")
+        return get_matrix_solenoid(sync_part=sync_part, length=length, B=B, charge=charge)
+
+    elif node_type is TiltLINAC:
+        angle = node.getTiltAngle()
+        if angle == 0:
+            return None
+        return get_matrix_tilt(sync_part=sync_part, angle=angle)
+
+    elif node_type is BaseRF_Gap:
+        E0TL = node.getParam("E0TL")
+        mode_phase = node.getParam("mode") * math.pi
+
+        cavity = node.getRF_Cavity()
+        frequency = cavity.getFrequency()
+        phase = cavity.getPhase() + mode_phase
+        amplitude = cavity.getAmp()
+
+        arrival_time = sync_part.time()
+        arrival_time_design = cavity.getDesignArrivalTime()
+
+        if node.isFirstRFGap():
+            if cavity.isDesignSetUp():
+                phase = math.fmod(frequency * (arrival_time - arrival_time_design) * 2.0 * math.pi + phase, 2.0 * math.pi)
+            else:
+                orbitFinalize("Run `trackDesign` first to initialize cavity phases.")
+        else:
+            phase = math.fmod(frequency * (arrival_time - arrival_time_design) * 2.0 * math.pi + phase,2.0 * math.pi)
+
+        node.setGapPhase(phase)
+
+        if amplitude == 0.0:
+            return None
+
+        return get_matrix_rf_gap(
+            sync_part=sync_part,
+            frequency=frequency,
+            E0TL=(E0TL * amplitude),
+            phase=phase,
+            charge=charge,
+        )
+
+    raise NotImplementedError(str(node))
