+++ py/orbit/envelope/meson.build
@@ -2,6 +2,7 @@ py_sources = files([
     '__init__.py',
     'envelope.py',
     'matrix.py',
+    'track.py',
     'utils.py'
 ])
